refactor init.lua to just define the public interface of the plugin. move plugin functionality to core module
